requirements.txt
Django>=1.10.6
six
django-environ
bumpversion
django-braces
python-dateutil
factory_boy
django-model-utils
monotonic
pypandoc